fix: resolve all build errors across capture and storage packages
capture:
- Move frameDuration to frame.go (shared, no build tag) to fix
  "undefined: frameDuration" on Windows builds
- Add Capturer interface to frame.go
- Add NewCapturer() factory to each platform file (windows, linux+x11,
  unsupported) so main.go doesn't hardcode a platform-specific constructor
- Add capture_unsupported.go (build !windows && !(linux && x11)) with
  stubs for queryWindowContextImpl and captureDisplayStub to allow
  compilation on plain Linux / CI environments
- Update main.go to call capture.NewCapturer instead of NewX11Capturer

storage:
- Fix pbkdf2.Key call for Go 1.24 API: new signature is
  Key(h func() Hash, password string, salt []byte, iter, keyLen int)
  ([]byte, error) - add string() cast on password, capture error return
- Fix retention.go: use int64(4096) in min() to match int64 size arg

+++ internal/storage/retention.go
@@ -214,7 +214,7 @@ func secureDeleteFile(path string) error {
 			f.Close()
 			return err
 		}
-		buf := make([]byte, min(4096, int(size)))
+		buf := make([]byte, min(int64(4096), size))
 		for i := range buf {
 			buf[i] = fillByte
 		}
